\name{fiboSeries-package}
\alias{fiboSeries-package}
\alias{fiboSeries}
\docType{package}
\title{
\packageTitle{fiboSeries}
}
\description{
\packageDescription{fiboSeries}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fiboSeries}
\packageIndices{fiboSeries}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
\packageAuthor{fiboSeries}

Maintainer: \packageMaintainer{fiboSeries}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per ~~
~~ line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
